// backend/src/schemas/auth.schema.ts
import { z } from 'zod';

export const registerSchema = z.object({
  fullName: z.string().min(2).max(100),
  email: z.string().email().max(255),
  password: z.string().min(8).max(72),
  companyName: z.string().min(2).max(100).optional(),
});
export type RegisterInput = z.infer<typeof registerSchema>;

export const loginSchema = z.object({
  email: z.string().email().max(255),
  password: z.string().min(1).max(72),
});
export type LoginInput = z.infer<typeof loginSchema>;